requirements.txt
click
pycairo